Fix infinite loop introduced in 11fd5e4784fa135cbdb64906cc628edb9e9d50ee

+++ libyara/re_lexer.c
@@ -1122,7 +1122,7 @@ YY_RULE_SETUP
   //  [abc0-9]
   //      ^- matching here
 
-  uint8_t c;
+  uint16_t c;
   uint8_t start = yytext[0];
   uint8_t end = yytext[2];
